feat: delete files from utils and create a unique utils file to integrate all functionalities
